backend/requirements.txt
amqp==5.3.1
bidict==0.23.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.2
MarkupSafe==3.0.2
numpy==2.2.4
pyrtlsdr==0.3.0
pyrtlsdrlib==0.0.3
python-engineio==4.11.2
python-socketio==5.12.1
scipy==1.15.2
setuptools==78.1.0
simple-websocket==1.1.0
tzdata==2025.2
vine==5.1.0
Werkzeug==3.1.3
wsproto==1.2.0
